+++ src/commands/here.php
@@ -49,6 +49,13 @@ if ( empty( $reset ) ) {
 	$here_dir = $plugins_dir;
 }
 
+// Due to a Docker bug, confirm Linux has proper permissions, otherwise output an error.
+$check_permissions = confirm_linux_has_file_permissions( [ $here_dir, $wp_dir, $themes_dir ] );
+
+if ( $check_permissions !== true ) {
+	echo red( $check_permissions );
+}
+
 $has_wp_config = dir_has_wp_config( $here_dir );
 $env_values    = [];